Extract job-handling code to rehearse package
