+++ pkg/rehearse/jobs_test.go
@@ -1,4 +1,4 @@
-package main
+package rehearse
 
 import (
 	"testing"
